refactor(routes): extract basic status routes
