+++ docs/ARCHITECTURE_PHASE_1_ROADMAP.md
@@ -120,6 +120,8 @@ Progress:
   preserving `from relaytv_app.routes import router`.
 - Extracted standalone `/health` into `app/relaytv_app/routes/health.py`.
 - Extracted standalone `/devices` into `app/relaytv_app/routes/devices.py`.
+- Extracted standalone `/discovery/status` and `/tv/status` into
+  `app/relaytv_app/routes/status.py`.
 
 ### M3: Extract Queue And Playback Routers
 
@@ -245,6 +247,7 @@ Add entries here as PRs land into `codex/architecture-phase-1`.
 | 2026-06-30 | local | `codex/architecture-phase-1` | Captured the public route inventory and added a route snapshot test before moving route code. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Begin M2 with low-risk router extraction. |
 | 2026-06-30 | local | `codex/architecture-phase-1` | Started M2 by converting `routes.py` to a package and extracting standalone `/health`. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Continue M2 with status/assets routes. |
 | 2026-06-30 | local | `codex/architecture-phase-1` | Continued M2 by extracting standalone `/devices`. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Continue M2 with status/assets routes. |
+| 2026-06-30 | local | `codex/architecture-phase-1` | Continued M2 by extracting `/discovery/status` and `/tv/status`. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Continue M2 with assets or app info. |
 
 ## Open Questions
